Weather API test: Wait for a bit before Mockito verifies if method was called
Sometimes the runnable posted to runOnServiceThread() might run a bit
faster than the calls made to the mocked weather provider, which causes
Mockito to throw an exception. Add a timeout to instruct Mockito to
wait for a bit before it verifies if the method was actually called.

Also removed an unused variable in CMWeatherManagerTest

TICKET: CYNGNOS-3043
Change-Id: I53b6541aeb571a6c83d1c26b6723d74c3661e8cc
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>